rename routers by routes
