+++ backend/matcha/routes/images.py
